+++ source/databricks/package/constants/basis_data_colname.py
@@ -12,12 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dataclasses import dataclass
 
-
-@dataclass
-class Metadata:
-    JobId: str
-    ResultId: str
-    ResultName: str
-    ResultPath: str
+class BasisDataColname:
+    energy_supplier_id = "ENERGYSUPPLIERID"
+    grid_area = "GRIDAREA"
+    in_grid_area = "FROMGRIDAREA"
+    metering_point_id = "METERINGPOINTID"
+    metering_point_type = "TYPEOFMP"
+    out_grid_area = "TOGRIDAREA"
+    quantity_prefix = "ENERGYQUANTITY"
+    "The column name prefix. The full quantity column names are suffixed with a number. E.g. 'ENERGYQUANTITY1'"
+    settlement_method = "SETTLEMENTMETHOD"
+    start_datetime = "STARTDATETIME"
+    valid_from = "VALIDFROM"
+    valid_to = "VALIDTO"