+++ source/dotnet/Services/Infrastructure/Processes/ProcessStepResultRepository.cs
@@ -69,14 +69,7 @@ public class ProcessStepResultRepository : IProcessStepResultRepository
 
     private async Task<ProcessStepResult> GetResultAsync(Guid batchId, string directory, TimeSeriesType timeSeriesType)
     {
-        var dataLakeFileClient =
-            await _dataLakeClient.GetDataLakeFileClientAsync(directory, ".json").ConfigureAwait(false);
-        if (dataLakeFileClient == null)
-        {
-            throw new InvalidOperationException($"Blob for batch with id={batchId} was not found.");
-        }
-
-        var resultStream = await dataLakeFileClient.OpenReadAsync(false).ConfigureAwait(false);
+        var resultStream = await _dataLakeClient.FindAndOpenFileAsync(directory, ".json").ConfigureAwait(false);
         var points = await _jsonNewlineSerializer.DeserializeAsync<ProcessResultPoint>(resultStream).ConfigureAwait(false);
 
         return MapToProcessStepResultDto(timeSeriesType, points);
@@ -90,9 +83,21 @@ public class ProcessStepResultRepository : IProcessStepResultRepository
                 point => new TimeSeriesPoint(
                     DateTimeOffset.Parse(point.quarter_time),
                     decimal.Parse(point.quantity, CultureInfo.InvariantCulture),
-                    point.quality))
+                    MapQuality(point.quality)))
             .ToList();
 
         return new ProcessStepResult(timeSeriesType, pointsDto.ToArray());
+    }
+
+    private static QuantityQuality MapQuality(string pointQuality)
+    {
+        return pointQuality switch
+        {
+            "measured" => QuantityQuality.Measured,
+            "calculated" => QuantityQuality.Calculated,
+            "estimated" => QuantityQuality.Estimated,
+            "Incomplete" => QuantityQuality.Incomplete,
+            _ =>throw new ArgumentException($"quality of unknown type:{pointQuality}"),
+        };
     }
 }